+++ tests/exploit-tests/beamsplitterReinitializationAttack.ts
@@ -7,11 +7,11 @@ import { SystemProgram } from "@solana/web3.js";
 import chai, { assert, expect } from "chai";
 
 import { enumLikeToString, generatePrismEtfAddress } from "../../src";
-import { coherenceHelper } from "../helper";
+import { coherenceHelper } from "../coherenceBeamsplitterTest";
 
 chai.use(chaiSolana);
 
-describe("beamsplitter reintialization attack", () => {
+export default function beamsplitterReinitializationAttack() {
   it("Reintialization attack should fail", async () => {
     const [initPrismEtFTx, prismEtfMint] =
       await coherenceHelper.sdk.initPrismEtf({
@@ -58,4 +58,4 @@ describe("beamsplitter reintialization attack", () => {
       "Signature verification failed"
     );
   });
-});
+}
