+++ tests/exploit-tests/index.ts
@@ -1,0 +1,7 @@
+import beamsplitterReinitializationAttack from "./beamsplitterReinitializationAttack";
+
+export default function exploitTests() {
+  describe("#EXPLOIT TESTS", () => {
+    beamsplitterReinitializationAttack();
+  });
+}
