+++ tests/stress-tests/index.ts
@@ -1,0 +1,7 @@
+import prismEtfSize from "./prismEtfSize";
+
+export default function stressTests() {
+  describe("#STRESS TESTS", () => {
+    prismEtfSize();
+  });
+}
